chore(magic-cloud-networking): update generated types and methods

+++ src/resources/magic-cloud-networking/catalog-syncs/catalog-syncs.ts
@@ -1,6 +1,7 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 import { APIResource } from '../../../resource';
+import { isRequestOptions } from '../../../core';
 import * as Core from '../../../core';
 import * as PrebuiltPoliciesAPI from './prebuilt-policies';
 import {
@@ -23,7 +24,7 @@ export class CatalogSyncs extends APIResource {
     params: CatalogSyncCreateParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<CatalogSyncCreateResponse> {
-    const { account_id, forwarded, ...body } = params;
+    const { account_id = this._client.accountId, forwarded, ...body } = params;
     return (
       this._client.post(`/accounts/${account_id}/magic/cloud/catalog-syncs`, {
         body,
@@ -41,7 +42,7 @@ export class CatalogSyncs extends APIResource {
     params: CatalogSyncUpdateParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<CatalogSyncUpdateResponse> {
-    const { account_id, ...body } = params;
+    const { account_id = this._client.accountId, ...body } = params;
     return (
       this._client.put(`/accounts/${account_id}/magic/cloud/catalog-syncs/${syncId}`, {
         body,
@@ -54,10 +55,20 @@ export class CatalogSyncs extends APIResource {
    * List Catalog Syncs (Closed Beta).
    */
   list(
-    params: CatalogSyncListParams,
+    params?: CatalogSyncListParams,
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<CatalogSyncListResponsesSinglePage, CatalogSyncListResponse>;
+  list(
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<CatalogSyncListResponsesSinglePage, CatalogSyncListResponse>;
+  list(
+    params: CatalogSyncListParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.PagePromise<CatalogSyncListResponsesSinglePage, CatalogSyncListResponse> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.list({}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return this._client.getAPIList(
       `/accounts/${account_id}/magic/cloud/catalog-syncs`,
       CatalogSyncListResponsesSinglePage,
@@ -70,10 +81,19 @@ export class CatalogSyncs extends APIResource {
    */
   delete(
     syncId: string,
-    params: CatalogSyncDeleteParams,
+    params?: CatalogSyncDeleteParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<CatalogSyncDeleteResponse>;
+  delete(syncId: string, options?: Core.RequestOptions): Core.APIPromise<CatalogSyncDeleteResponse>;
+  delete(
+    syncId: string,
+    params: CatalogSyncDeleteParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<CatalogSyncDeleteResponse> {
-    const { account_id, delete_destination } = params;
+    if (isRequestOptions(params)) {
+      return this.delete(syncId, {}, params);
+    }
+    const { account_id = this._client.accountId, delete_destination } = params;
     return (
       this._client.delete(`/accounts/${account_id}/magic/cloud/catalog-syncs/${syncId}`, {
         query: { delete_destination },
@@ -90,7 +110,7 @@ export class CatalogSyncs extends APIResource {
     params: CatalogSyncEditParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<CatalogSyncEditResponse> {
-    const { account_id, ...body } = params;
+    const { account_id = this._client.accountId, ...body } = params;
     return (
       this._client.patch(`/accounts/${account_id}/magic/cloud/catalog-syncs/${syncId}`, {
         body,
@@ -104,10 +124,19 @@ export class CatalogSyncs extends APIResource {
    */
   get(
     syncId: string,
-    params: CatalogSyncGetParams,
+    params?: CatalogSyncGetParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<CatalogSyncGetResponse>;
+  get(syncId: string, options?: Core.RequestOptions): Core.APIPromise<CatalogSyncGetResponse>;
+  get(
+    syncId: string,
+    params: CatalogSyncGetParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<CatalogSyncGetResponse> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.get(syncId, {}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return (
       this._client.get(
         `/accounts/${account_id}/magic/cloud/catalog-syncs/${syncId}`,
@@ -122,10 +151,19 @@ export class CatalogSyncs extends APIResource {
    */
   refresh(
     syncId: string,
-    params: CatalogSyncRefreshParams,
+    params?: CatalogSyncRefreshParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<CatalogSyncRefreshResponse>;
+  refresh(syncId: string, options?: Core.RequestOptions): Core.APIPromise<CatalogSyncRefreshResponse>;
+  refresh(
+    syncId: string,
+    params: CatalogSyncRefreshParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<CatalogSyncRefreshResponse> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.refresh(syncId, {}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return (
       this._client.post(
         `/accounts/${account_id}/magic/cloud/catalog-syncs/${syncId}/refresh`,
@@ -1222,7 +1260,7 @@ export interface CatalogSyncCreateParams {
   /**
    * Path param
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Body param
@@ -1259,7 +1297,7 @@ export interface CatalogSyncUpdateParams {
   /**
    * Path param
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Body param
@@ -1283,14 +1321,14 @@ export interface CatalogSyncUpdateParams {
 }
 
 export interface CatalogSyncListParams {
-  account_id: string;
+  account_id?: string;
 }
 
 export interface CatalogSyncDeleteParams {
   /**
    * Path param
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Query param
@@ -1302,7 +1340,7 @@ export interface CatalogSyncEditParams {
   /**
    * Path param
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Body param
@@ -1326,11 +1364,11 @@ export interface CatalogSyncEditParams {
 }
 
 export interface CatalogSyncGetParams {
-  account_id: string;
+  account_id?: string;
 }
 
 export interface CatalogSyncRefreshParams {
-  account_id: string;
+  account_id?: string;
 }
 
 CatalogSyncs.CatalogSyncListResponsesSinglePage = CatalogSyncListResponsesSinglePage;
